Extract ToolbarContent into their own files.
- This will set up to allow exploring having one .toolbar in the app, so placement may be more managable.
